@@ -105,7 +105,7 @@ if __name__=='__main__':
     # The main I/O work starts now
     #
 
-    print 'Reading data...\n'
+    print('Reading data...\n')
     data = read(path_data, format='sac',
         event_id=event_name,
         tags=['units:cm', 'type:velocity']) 
@@ -116,14 +116,14 @@ if __name__=='__main__':
     origin = data.get_origins()[0]
 
 
-    print 'Processing data...\n'
+    print('Processing data...\n')
     data_bw = data.map(process_bw)
     data_sw = data.map(process_sw)
 
-    print 'Reading Green''s functions...\n'
+    print('Reading Green''s functions...\n')
     greens = download_greens_tensors(stations, origin, model)
 
-    print 'Processing Greens functions...\n'
+    print('Processing Greens functions...\n')
     greens.convolve(wavelet)
     greens_bw = greens.map(process_bw)
     greens_sw = greens.map(process_sw)
@@ -133,12 +133,12 @@ if __name__=='__main__':
     # The main computational work starts nows
     #
 
-    print 'Evaluating body wave misfit...\n'
+    print('Evaluating body wave misfit...\n')
 
     results_bw = grid_search(
         data_bw, greens_bw, misfit_bw, origin, sources)
 
-    print 'Evaluating surface wave misfit...\n'
+    print('Evaluating surface wave misfit...\n')
 
     results_sw = grid_search(
         data_sw, greens_sw, misfit_sw, origin, sources)
@@ -151,7 +151,7 @@ if __name__=='__main__':
     # Saving results
     #
 
-    print 'Saving results...\n'
+    print('Saving results...\n')
 
     plot_data_greens(event_name+'.png', 
         data_bw, data_sw, greens_bw, greens_sw, process_bw, process_sw, 
@@ -161,5 +161,5 @@ if __name__=='__main__':
 
     #grid.save(event_name+'.h5', {'misfit': results})
 
-    print 'Finished\n'
+    print('Finished\n')
 
