@@ -8,10 +8,15 @@ import time
 import numpy as np
 import obspy
 import re
-import urllib
 import uuid
 import warnings
 import zipfile
+
+
+try:
+    from urllib import URLopener
+except ImportError:
+    from urllib.request import URLopener
 
 
 class AttribDict(obspy.core.util.attribdict.AttribDict):
@@ -72,7 +77,7 @@ def timer(func):
 
         if kwargs.get('verbose', True):
             _elapsed_time = time.time() - start_time
-            print '  Elapsed time (s): %f\n' % _elapsed_time
+            print('  Elapsed time (s): %f\n' % _elapsed_time)
 
         return output
 
@@ -100,7 +105,7 @@ def timer(func):
             start_time = time.time()
             output = func(*args, **kwargs)
             elapsed_time = time.time() - start_time
-            print '  Elapsed time (s): %f\n' % elapsed_time
+            print('  Elapsed time (s): %f\n' % elapsed_time)
             return output
         else:
             return func(*args, **kwargs)
@@ -135,8 +140,8 @@ def warn(*args, **kwargs):
 
 @retry(Exception, tries=4, delay=2, backoff=2)
 def urlopen_with_retry(url, filename):
-    download = urllib.URLopener()
-    download.retrieve(url, filename)
+    opener = URLopener()
+    opener.retrieve(url, filename)
 
 
 def url2uuid(url):
