@@ -6,18 +6,18 @@ if __name__ == '__main__':
 
     for key in ['readers', 'greens_tensor_clients']:
 
-        print key.upper()
+        print(key.upper())
 
         for entry_point in iter_entry_points(key):
 
             try:
-                print '    %s' % entry_point.name
+                print('    %s' % entry_point.name)
             except:
                 raise Exception
 
             entry_point.load()
 
-        print ''
+        print('')
 
 
 
