@@ -52,9 +52,9 @@ def check_module(name):
         excluded += [name]
     except Exception as e:
         if verbosity > 0:
-            print name.upper()
-            print e
-            print '\n'
+            print(name.upper())
+            print(e)
+            print('\n')
         failed += [name]
     else:
         succeeded += [name]
@@ -65,33 +65,33 @@ if __name__ == '__main__':
     check_package('mtuq')
 
     # disply results
-    print '\n'
-    print '%d modules were excluded' % len(excluded)
-    print '\n'
-    print '%d modules imported successfully' % len(succeeded)
-    print '\n'
-    print '%d modules failed to import' % len(failed)
-    print '\n'
+    print('\n')
+    print('%d modules were excluded' % len(excluded))
+    print('\n')
+    print('%d modules imported successfully' % len(succeeded))
+    print('\n')
+    print('%d modules failed to import' % len(failed))
+    print('\n')
 
     if verbosity > 1:
         if excluded:
-            print 'The following modules were excluded:'
-            for name in excluded: print '  %s' % name
-            print '\n'
+            print('The following modules were excluded:')
+            for name in excluded: print('  %s' % name)
+            print('\n')
 
         if succeeded:
-            print 'The following modules imported successfully:'
-            for name in succeeded: print '  %s' % name
-            print '\n'
+            print('The following modules imported successfully:')
+            for name in succeeded: print('  %s' % name)
+            print('\n')
 
         if failed:
-            print 'The following modules failed to import:'
-            for name in failed: print '  %s' % name
-            print '\n'
+            print('The following modules failed to import:')
+            for name in failed: print('  %s' % name)
+            print('\n')
         
     if failed:
         sys.exit(-1)
     else:
-        print 'SUCCESS\n'
+        print('SUCCESS\n')
 
 
