@@ -97,7 +97,7 @@ if __name__=='__main__':
     # The main I/O work starts now
     #
 
-    print 'Reading data...\n'
+    print('Reading data...\n')
     data = read(path_data, format='sac', 
         event_id=event_name,
         tags=['units:cm', 'type:velocity']) 
@@ -112,15 +112,15 @@ if __name__=='__main__':
         origins += [origin.copy()]
         setattr(origins[-1], 'depth_in_m', depth)
 
-    print 'Processing data...\n'
+    print('Processing data...\n')
     data_bw = data.map(process_bw)
     data_sw = data.map(process_sw)
 
-    print 'Reading Green''s functions...\n'
+    print('Reading Green''s functions...\n')
     db = open_db(path_greens, format='FK', model=model)
     greens = db.get_greens_tensors(stations, origins)
 
-    print 'Processing Greens functions...\n'
+    print('Processing Greens functions...\n')
     greens.convolve(wavelet)
     greens_bw = greens.map(process_bw)
     greens_sw = greens.map(process_sw)
@@ -130,12 +130,12 @@ if __name__=='__main__':
     # The main computational work starts now
     #
 
-    print 'Evaluating body wave misfit...\n'
+    print('Evaluating body wave misfit...\n')
 
     results_bw = grid_search(
         data_bw, greens_bw, misfit_bw, origins, sources, verbose=False)
 
-    print 'Evaluating surface wave misfit...\n'
+    print('Evaluating surface wave misfit...\n')
 
     results_sw = grid_search(
         data_sw, greens_sw, misfit_sw, origins, sources, verbose=False)
@@ -152,5 +152,5 @@ if __name__=='__main__':
     if run_checks:
         pass
 
-    print 'SUCCESS\n'
+    print('SUCCESS\n')
 
