@@ -67,7 +67,7 @@ def read(path='', format='', **kwargs):
 # methods
 #
 
-import copy_reg
+import copyreg
 import types
 
 def _pickle_method(method):
@@ -87,7 +87,7 @@ def _unpickle_method(func_name, obj, cls):
             break
     return func.__get__(obj, cls)
 
-copy_reg.pickle(types.MethodType, _pickle_method, _unpickle_method)
+copyreg.pickle(types.MethodType, _pickle_method, _unpickle_method)
 
 
 
