@@ -1,3 +1,3 @@
 
-from base import GreensTensor, GreensTensorList
+from mtuq.greens_tensor.base import GreensTensor, GreensTensorList
 
